french comment order controller

+++ app/controllers/addressController.js
@@ -47,10 +47,7 @@ const addressController = {
             lastname_address: req.body.lastNameAddress,
             user_id: req.body.userId
         };
-        console.log('newaddressdata :', newAddressData)
         const newAddress = await Address.create(newAddressData);
-        console.log('newaddress :', newAddress)
-
         const theAddressUser = await Address.findOne({
             where: { id: newAddress.id },
             attributes: {
@@ -62,7 +59,7 @@ const addressController = {
 
     },
 
-    //on modifie un enregistrement on fonction de son id
+    // modification d'une address en fonction de son id
     updateById: async (req, res) => {
         const { id } = req.params;
         const newAddressData = {
@@ -77,24 +74,31 @@ const addressController = {
             user_id: req.body.userId
         };
         console.log('data renvoyée:', newAddressData)
-        //const oldUser = await Address.findOne({where:{id}});
+
         await Address.update({ ...newAddressData }, { where: { id } })
-        const newUser = await Address.findByPk(id)
-        res.json(newUser)
+        const newAddress = await Address.findByPk(id)
+        res.json(newAddress)
     },
+
+    // Suppression d'une addresse en fonction de son id
     deleteById: async (req, res) => {
-        try {
+       
             const { id } = req.params;
-            console.log('id:', id)
+            const address = await Address.findByPk(id);
+            if (!address) {
+                res.status(400).json(`l'addresse avec l'id ${id} n'existe pas et ne peut donc pas être supprimé`);
+                return next();
+            }
+
             const user = await Address.findByPk(id);
             Address.destroy({ where: { id } })
-            //await user.destroy();
-            res.json(`l'addresse avec l'id ${id} est bien supprimé`)
+
+            const addressExist = await Address.findByPk(id);
+            if (addressExist) {
+                res.status(400).json(`l'addresse avec l'id ${id} n'as pas était supprimé`);
+            };
+            res.json(`l'addresse avec l'id ${id} est bien supprimé`);
         }
-        catch {
-            res.status(400).json(`l'addresse avec l'id ${id} n'a pas pu être supprimé ou n'existe pas`)
-        }
-    }
 
 }
 module.exports = addressController;
